Revert "Fixes #123. App crashes when `google-service.json` is missing. (#127)"
This reverts commit 2e5f9b32454c67436a972cedd01ecfe28801f6da.
