+++ app/build.gradle
@@ -89,7 +89,6 @@ dependencies {
 	implementation(project(":screen:ui-bookmarks"))
 
 	implementation Firebase.dynamicLink
-	implementation(project(":features:article:dynamic-links-handler"))
 }
 
 if (file("google-services.json").exists()) {
