+++ features/article/dynamic-links-handler/src/main/AndroidManifest.xml
@@ -1,2 +1,2 @@
 <?xml version="1.0" encoding="utf-8"?>
-<manifest package="kasem.sm.dynamic_links_handler"/>
+<manifest package="com.amaan.dynamic_links_handler"/>
